site/trendspedia/requirements.txt
Django==1.4.3
Fabric==1.4.3
MySQL-python==1.2.4b5
amqp==1.0.6
anyjson==0.3.3
argparse==1.2.1
beautifulsoup4==4.1.3
billiard==2.7.3.19
celery==3.0.13
coverage==3.6
django-celery==3.0.11
django-coverage==1.2.2
django-discover-runner==0.3
django-mongodb-engine==0.4.0
django-social-auth==0.7.22
djangorestframework==2.0.0
djangotoolbox==0.9.2
httplib2==0.8
importlib==1.0.2
kombu==2.5.4
mongoengine==0.7.8
oauth2==1.5.211
oauthlib==0.4.0
ordereddict==1.1
paramiko==1.9.0
pika==0.9.8
pycrypto==2.6
pymongo==2.4.1
python-dateutil==1.5
python-openid==2.2.5
pytz==2012j
requests==1.2.0
requests-oauth==0.4.1
requests-oauthlib==0.3.0
simplejson==3.1.3
ssh==1.8.0
twython==2.7.3
uWSGI==1.3
wsgiref==0.1.2
